requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
python-multipart
supabase
postgrest
langchain
langchain-community
langchain-openai
pypdf
tiktoken
fastapi-mail
httpx
cloudinary
pylovepdf
langchain
langchain-groq
langgraph
tavily-python
google-generativeai
langchain-google-genai
langchain-text-splitters